moved zcl-reload test back to zcl-loader.test.js
